maildir: move functions
