+++ debug/graphviz.c
@@ -37,12 +37,13 @@
 #include "conn/lib.h"
 #include "lib.h"
 #include "compmbox/lib.h"
-#include "maildir/lib.h"
 #include "mbox/lib.h"
 #include "nntp/lib.h"
 #include "notmuch/lib.h"
 #include "context.h"
 #include "imap/private.h"
+#include "maildir/edata.h"
+#include "maildir/mdata.h"
 #include "maildir/private.h"
 #include "mutt_globals.h"
 #include "notmuch/private.h"
