+++ maildir/private.h
@@ -28,57 +28,15 @@
 #include <stdio.h>
 #include <sys/types.h>
 #include <time.h>
+#include "sequence.h"
 
 struct Account;
 struct Buffer;
 struct Email;
 struct Mailbox;
+struct Maildir;
 struct Message;
 struct Progress;
-
-/**
- * struct MaildirEmailData - Maildir-specific Email data - @extends Email
- */
-struct MaildirEmailData
-{
-  char *maildir_flags; ///< Unknown Maildir flags
-};
-
-/**
- * struct MaildirMboxData - Maildir-specific Mailbox data - @extends Mailbox
- */
-struct MaildirMboxData
-{
-  struct timespec mtime_cur;
-  mode_t mh_umask;
-};
-
-/**
- * struct Maildir - A Maildir mailbox
- */
-struct Maildir
-{
-  struct Email *email;
-  char *canon_fname;
-  bool header_parsed : 1;
-  ino_t inode;
-  struct Maildir *next;
-};
-
-typedef uint8_t MhSeqFlags;     ///< Flags, e.g. #MH_SEQ_UNSEEN
-#define MH_SEQ_NO_FLAGS      0  ///< No flags are set
-#define MH_SEQ_UNSEEN  (1 << 0) ///< Email hasn't been read
-#define MH_SEQ_REPLIED (1 << 1) ///< Email has been replied to
-#define MH_SEQ_FLAGGED (1 << 2) ///< Email is flagged
-
-/**
- * struct MhSequences - Set of MH sequence numbers
- */
-struct MhSequences
-{
-  int max;           ///< Number of flags stored
-  MhSeqFlags *flags; ///< Flags for each email
-};
 
 extern bool  C_CheckNew;
 extern bool  C_MaildirCheckCur;
@@ -104,23 +62,25 @@ int             mh_msg_save_hcache (struct Mailbox *m, struct Email *e);
 /* Maildir/MH shared functions */
 void                    maildir_canon_filename (struct Buffer *dest, const char *src);
 void                    maildir_delayed_parsing(struct Mailbox *m, struct Maildir **md, struct Progress *progress);
+void                    maildir_free           (struct Maildir **md);
 size_t                  maildir_hcache_keylen  (const char *fn);
-struct MaildirMboxData *maildir_mdata_get      (struct Mailbox *m);
 int                     maildir_mh_open_message(struct Mailbox *m, struct Message *msg, int msgno, bool is_maildir);
 int                     maildir_move_to_mailbox(struct Mailbox *m, struct Maildir **ptr);
 int                     maildir_parse_dir      (struct Mailbox *m, struct Maildir ***last, const char *subdir, int *count, struct Progress *progress);
+void                    maildir_update_mtime   (struct Mailbox *m);
+struct Maildir *        maildir_sort           (struct Maildir *list, size_t len, int (*cmp)(struct Maildir *, struct Maildir *));
+int                     md_cmp_inode           (struct Maildir *a, struct Maildir *b);
+int                     md_cmp_path            (struct Maildir *a, struct Maildir *b);
 int                     md_commit_message      (struct Mailbox *m, struct Message *msg, struct Email *e);
 int                     mh_commit_msg          (struct Mailbox *m, struct Message *msg, struct Email *e, bool updseq);
 int                     mh_mkstemp             (struct Mailbox *m, FILE **fp, char **tgt);
 int                     mh_read_dir            (struct Mailbox *m, const char *subdir);
-int                     mh_read_sequences      (struct MhSequences *mhs, const char *path);
-MhSeqFlags              mhs_check              (struct MhSequences *mhs, int i);
-void                    mhs_sequences_free     (struct MhSequences *mhs);
 MhSeqFlags              mhs_set                (struct MhSequences *mhs, int i, MhSeqFlags f);
+void                    mh_sort_natural        (struct Mailbox *m, struct Maildir **md);
 mode_t                  mh_umask               (struct Mailbox *m);
 void                    mh_update_maildir      (struct Maildir *md, struct MhSequences *mhs);
-void                    mh_update_sequences    (struct Mailbox *m);
 bool                    mh_valid_message       (const char *s);
+struct Maildir *        skip_duplicates        (struct Maildir *p, struct Maildir **last);
 
 int mh_sync_message(struct Mailbox *m, int msgno);
 int maildir_sync_message(struct Mailbox *m, int msgno);
